Format & Lint

+++ contracts/mock/MockParaswapTransferProxy.sol
@@ -8,12 +8,7 @@ import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 contract MockParaswapTransferProxy {
     using SafeERC20 for IERC20;
 
-    function safeTransferFrom(
-        address from,
-        address to,
-        address token,
-        uint amount
-    ) external {
+    function safeTransferFrom(address from, address to, address token, uint256 amount) external {
         IERC20(token).safeTransferFrom(from, to, amount);
     }
 }
